exported guestCart helper functions to guestHelper.js
