+++ client/app.js
@@ -2,9 +2,8 @@ import React from 'react'
 
 import {Navbar} from './components'
 import Routes from './routes'
-let cartList = {}
+
 const App = () => {
-  localStorage.setItem('guestCart', JSON.stringify(cartList))
   return (
     <div>
       <Navbar />
