+++ client/store/cart.js
@@ -2,6 +2,7 @@
 /* eslint-disable complexity */
 import axios from 'axios'
 import history from '../history'
+import {removeFromGuestCart, updateQuantityGuestCart} from './guestHelper'
 
 const SET_CART = 'SET_CART'
 const ADD_TO_CART = 'ADD_TO_CART'
@@ -76,23 +77,7 @@ export const removeProduct = (productId, userId = 0) => {
     }
   }
 }
-// helper function to remove items from localStorage
-function removeFromGuestCart(productId) {
-  const guestCart = JSON.parse(localStorage.getItem('guestCart'))
-  const removedProduct = guestCart[productId]
-  delete guestCart[productId]
-  localStorage.setItem('guestCart', JSON.stringify(guestCart))
-  console.log('remove ----->', JSON.parse(localStorage.getItem('guestCart')))
-  return removedProduct
-}
 
-// helper function to update quantity from localStorage
-function updateQuantityGuestCart(productId, quantity) {
-  const guestCart = JSON.parse(localStorage.getItem('guestCart'))
-  guestCart[productId].quantity = quantity
-  localStorage.setItem('guestCart', JSON.stringify(guestCart))
-  console.log('update ---->', JSON.parse(localStorage.getItem('guestCart')))
-}
 export const updateQuantityThunk = (productId, userId = 0, quantity) => {
   return async dispatch => {
     try {
